Move mobile search button up one component and hide download text on mobile
